Remove the logic which rethrows NoClassDefFound as ClassNotFoundException entirely. Instead, wrap all Exceptions thrown in mustBeLoaded() in a NoClassDefFoundError.
What we are trying to achieve, spec-wise:
1) If Class.forName() fails because the class loader it invokes throws
ClassNotFoundException, Class.forName() should also throw
ClassNotFoundException.
2) If the VM tries to load a class and the attempt fails, a
NoClassDefFoundError should be thrown.

We interpret this as follows:
1) Class.forName() does nothing special with whatever is thrown by
loadClass(). So if loadClass() would throw for example a
NullPointerException this too would be propagated to the caller.
2) When Mika function mustBeLoaded() returns CLASS_LOADING_FAILED, there
will always be an Error pending; if namedClassMustBeLoaded() throws an
Error then this is passed un unchanged, if it throws an Exception then
this will be wrapped in a NoClassDefFoundError.