Heartbeat now calls a native method nativesleep() between each "beat". This method also detects any clock "jumps" and accumulates them in a variable which can be read using Heartbeat.getTimeOffset(). This is used by java.util.TimerThread to compensate for such timer jumps when calculating the next execution time for a task.
If the system clock jumps backwards, an attempt is made to notify all
threads which are executing a pthread_cond_timedwait(). This seems not
to work, so in fact this patch only works correctly when HAVE_TIMEDWAIT
is not defined; in this case an internal timing mechanism is used
instead of pthread_cond_timedwait().

Note that with this change we introduce a deviation from the observed
behaviour of Sun's HotSpot VM, at least on Linux: on that platform a
change to the system clock results in timers expiring very rapidly
(clock advance) or not at all until the system clock "catches up" with
its previous value (clock retard). This behaviour is only a minor
nuisance on desktop or server systems, in which changes to the system
clock are rare and/or gradual (e.g. ntpd); however it is undesirably on
embedded systems where the system clock may suddenly change as a result
of user actions or of an automatic action such as setting the time from
a remote server.

This changeset also includes some "tidying up" in this area, for example
removing the unused doubly-linked list fields from x_Thread and ensuring
that only the needed fields are defined according to whether
HAVE_TIMEDWAIT is set.