Allow GC triggered by a call to heap_request to be postponed until we reach a safe point, thereby allowing heap_request to be called in both safe and unsafe states.
In internString, if canonical String instance is retrieved set it
O_BLACK (and clear O_GARBAGE): it may be that this String instance was
previously unreachable and that a sweep is already in progress.

Cleanup: eliminate USE_INTERNED_STRING_HASHTABLE option from strings.c.

Move definition of PIGS_MIGHT_FLY from collector.c to wonka.h, so that
it is also visible in strings.c.